[FIX] stock_vertical_lift: Fix handling of refresh notifications
- Delete channel only when unmounting vertical lift operation form view
- Use soft reload instead of hard reload
- Bind handling of refresh notification through subscribe instead
  of event listener
- Broadcast refresh notification on plain channel string (ie channel name)
  instead of composite channel (ie partner_id, channel tuple), since
  external API calls may be executed by another user than the one
  interacting with the vertical lift.

+++ stock_vertical_lift/static/src/js/vertical_lift.esm.js
@@ -68,26 +68,26 @@ patch(FormController.prototype, {
         this.busService = useService("bus_service");
         if (this.props.resModel.startsWith("vertical.lift.operation.")) {
             this.busService.addChannel("notify_vertical_lift_screen");
-            this.busService.addEventListener("notification", (notifications) => {
-                notifications.forEach(([channel, message]) => {
-                    if (
-                        channel === "notify_vertical_lift_screen" &&
-                        message.action === "refresh"
-                    ) {
-                        this.vlift_bus_action_refresh(message.params);
-                    }
-                });
+            this.busService.subscribe("notification", (payload) => {
+                if (payload.action === "refresh") {
+                    this.vlift_bus_action_refresh(payload.params);
+                }
             });
         }
 
         onWillUnmount(() => {
-            this.busService.deleteChannel("notify_vertical_lift_screen");
+            if (this.props.resModel.startsWith("vertical.lift.operation.")) {
+                this.busService.deleteChannel("notify_vertical_lift_screen");
+            }
         });
     },
 
     vlift_bus_action_refresh(params) {
         if (params.id === this.props.resId && params.model === this.props.resModel) {
-            this.model.root.load();
+            this.env.services.action.doAction({
+                type: "ir.actions.client",
+                tag: "soft_reload",
+            });
         }
     },
 });
